Added client/server disconnect 'handshake'
Fixed bug where reconnected to the same server after disconnecting (without restarting application) would not allow the client to receive data. The server was able to send/receive data, the client was able to send files, but the client would not receive any requested files/folders, and any changes to the remote workspace would not update the remote file tree for the server.

This was solved by having the client alert the server that it intends to close the connection and nesting the client's input stream in an infinite loop that only closes when the client is closed.