properly setup maven modules and pom files
also added FileConflictListener to the Client module
This listener is meant to intercept write messages being sent to the server to determine if the message sent can potentially cause file conflicts due to naming issues on the receiving end (for file overwrites etc). Still needs to be implemented for actions on local file space.

I also moved the searchDirectory() static method inside of FSRemoteFileTreeUtil to FSUtil